Fixed numerous little issues that prevernted textproc from working.
Mostly involved resolving issues created when level 2 headings were
created early, and the rest of the code was not expecting them.
Also involved modifying some code that created lists and then adding
links, when other code did not need to create links; so created a
separate macro that did not create links.